backup_previous_versions/requirements.txt
datasets
scikit-learn
torch
transformers
ray
google
protobuf
accelerate
wandb
peft
bitsandbytes
sentencepiece
sentence-transformers
python-dotenv
wheel
setuptools
matplotlib
jsonargparse
setfit
# FLASH_ATTENTION_SKIP_CUDA_BUILD=TRUE pip install flash-attn --no-build-isolation